Split off PngTerrainGenerator, SimplexNoise units
