+++ src/openrct2-ui/windows/MapGen.cpp
@@ -26,6 +26,7 @@
 #include <openrct2/windows/Intent.h>
 #include <openrct2/world/Map.h>
 #include <openrct2/world/map_generator/MapGen.h>
+#include <openrct2/world/map_generator/PngTerrainGenerator.h>
 
 namespace OpenRCT2::Ui::Windows
 {
