+++ src/openrct2/libopenrct2.vcxproj
@@ -626,6 +626,8 @@
     <ClInclude Include="world\TilePointerIndex.hpp" />
     <ClInclude Include="world\map_generator\MapGen.h" />
     <ClInclude Include="world\map_generator\MapHelpers.h" />
+    <ClInclude Include="world\map_generator\PngTerrainGenerator.h" />
+    <ClInclude Include="world\map_generator\SimplexNoise.h" />
     <ClInclude Include="world\tile_element\BannerElement.h" />
     <ClInclude Include="world\tile_element\EntranceElement.h" />
     <ClInclude Include="world\tile_element\LargeSceneryElement.h" />
@@ -1122,6 +1124,8 @@
     <ClCompile Include="world\TileInspector.cpp" />
     <ClCompile Include="world\map_generator\MapGen.cpp" />
     <ClCompile Include="world\map_generator\MapHelpers.cpp" />
+    <ClCompile Include="world\map_generator\PngTerrainGenerator.cpp" />
+    <ClCompile Include="world\map_generator\SimplexNoise.cpp" />
     <ClCompile Include="world\tile_element\BannerElement.cpp" />
     <ClCompile Include="world\tile_element\EntranceElement.cpp" />
     <ClCompile Include="world\tile_element\LargeSceneryElement.cpp" />
